requirements.txt
google-generativeai==0.3.2
python-dotenv==1.0.1
# Editable install with no version control (termtel==0.1.0)
-e /home/screenager/Projects/CLI/termtel

